Add default metrics to 'Evaluator'
