+++ src/gluonts/ev/api.py
@@ -148,53 +148,3 @@ class DerivedMetricEvaluator(MetricEvaluator):
     def reset(self) -> None:
         for metric_evaluator in self.metrics.values():
             metric_evaluator.reset()
-
-
-class Evaluator:
-    def __init__(self) -> None:
-        # TODO: better naming!
-        self.metric_evaluators: dict[str, MetricEvaluator] = dict()
-
-    def add_metric(self, metrics: Metric, axis: Optional[int] = None) -> None:
-        self.add_metrics([metrics], axis)
-
-    def add_metrics(
-        self, metrics: Collection[Metric], axis: Optional[int] = None
-    ) -> None:
-        for metric in metrics:
-            metric_evaluator = metric(axis=axis)
-            metric_name = f"{metric.__class__.__name__}[axis={axis}]"
-            self.metric_evaluators[metric_name] = metric_evaluator
-
-    def get_required_quantile_levels(self, test_data: TestData):
-        data_probe = DataProbe(test_data)
-        for metric_evaluator in self.metric_evaluators.values():
-            metric_evaluator.reset()
-            metric_evaluator.update(data_probe)
-            metric_evaluator.get()
-        return data_probe.required_quantile_forecasts
-
-    def evaluate(
-        self, test_data: TestData, forecasts: Iterator[Forecast]
-    ) -> EvalData:
-        quantile_levels = self.get_required_quantile_levels(test_data)
-
-        batches = gather_inputs(
-            test_data=test_data,
-            forecasts=forecasts,
-            quantile_levels=quantile_levels,
-        )
-
-        # only NumPy arrays are used from here on
-
-        for metric_evaluator in self.metric_evaluators.values():
-            metric_evaluator.reset()
-
-        for data in batches:
-            for metric_evaluator in self.metric_evaluators.values():
-                metric_evaluator.update(data)
-
-        result = dict()
-        for metric_name, metric_evaluator in self.metric_evaluators.items():
-            result[metric_name] = metric_evaluator.get()
-        return result
