+++ src/gluonts/ev/usage_example.py
@@ -11,17 +11,10 @@
 # express or implied. See the License for the specific language governing
 # permissions and limitations under the License.
 
-from toolz import take
 
-from gluonts.ev.api import Evaluator
+from toolz import take
+from gluonts.ev.evaluator import Evaluator
 from gluonts.dataset.split import TestTemplate, OffsetSplitter
-from gluonts.ev.metrics import (
-    MSE,
-    MSIS,
-    NRMSE,
-    QuantileLoss,
-    WeightedQuantileLoss,
-)
 from gluonts.model.npts import NPTSPredictor
 from gluonts.dataset.repository.datasets import get_dataset
 
@@ -46,27 +39,9 @@ forecast_it = predictor.predict(dataset=test_data.input, num_samples=100)
 # --- EVALUATION STARTS HERE ---
 
 evaluator = Evaluator()
-evaluator.add_metrics([MSE(), MSIS(), WeightedQuantileLoss()], axis=1)
-evaluator.add_metric(NRMSE())
-evaluator.add_metric(QuantileLoss(q=0.9), axis=0)
+evaluator.add_default_metrics()
 
 res = evaluator.evaluate(test_data, forecast_it)
 
 for name, value in res.items():
     print(f"\n{name}: {value}")
-
-
-"""
-
-MSE[axis=1]: [1.52131758e+02 6.88867542e+01 3.00020833e-01 1.15193784e+03
- 3.15931108e+02 2.63484380e+03 2.99962083e+00 4.66641007e+03
- 2.05939438e+03 4.50878403e+03]
-
-MSIS[axis=1]: [12.08145374  4.83207589  0.62075673  3.42150118  4.37868274  3.2297932
-  3.44080163  3.83119893 10.97729275  5.0347862 ]
-
-NRMSE[axis=None]: 0.14527278126172422
-
-QuantileLossSum[axis=0]: [ 58.3  54.5  33.7  35.   43.1  31.8  31.2  47.1  52.9  55.1  37.7  31.7
-  44.7  60.7 176.7 160.4  54.1  29.5  52.6  53.3  47.2  57.6  62.5 134.5]
-"""
